requirements.txt
python==3.8.5
pytorch==1.7.0
torchvision==0.8.1
imageio==2.9.0
opencv==4.0.1
opencv-python==4.4.0.44
cython==0.29.21
matplotlib==3.3.1
numpy==1.19.1
